personal data index page only viewed in datas and  personal detail index page view data and edit data
