+++ src/App.js
@@ -50,6 +50,8 @@ function App() {
           <Route path="/loans/:loanName" element={<SecuredLoansDetails />} />
           <Route path="/showresult" element={<ShowResult />} />
           <Route path="/emicalculator" element={<EmiCalculator/>}/>
+          <Route path="/personaldetail" element={<PersonalDetail/>}/>
+          <Route path="/loandatas" element={<PersonalDatas/>}/>
         </Routes>
       </BrowserRouter>
     </div>
