+++ src/Finance/Category/index.js
@@ -24,7 +24,7 @@ export default function Category (){
         <Link className="nav-link" to={'/emicalculator'}>EMI Calulator</Link>
       </li>
       <li className="nav-item">
-        <Link className="nav-link" to={'/personaldatas'}>Your Datas</Link>
+        <Link className="nav-link" to={'/loandatas/:loanName'}>Loan Datas</Link>
       </li>
       <li class="nav-item dropdown">
         <Link class="nav-link dropdown-toggle" href="#" id="navbarDropdownMenuLink" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
