+++ src/Finance/FinanceDetail/SecuredLoansDetails.js
@@ -80,22 +80,17 @@ const handleSetLoanData = async () => {
 
  
   
-  // if(filteredInputNames.some(name => !securedLoansInfo[name])){
-  //   alert("pls fill the inputs")
-  // }
+  
    if(securedLoansInfo.propertyStatus == "owned" && securedLoansInfo.CibilIssue == "no" && securedLoansInfo.monthlyNetIncome >= 25000 && elgAmount ) {
-    // await addDoc(collection(db, "securedLoans"), {
-    //         ...securedLoansInfo,uId: userdata.uid,loanType:loanName,grade:"A"});
+   
     alert("grade A")
   }
   else if(securedLoansInfo.CibilIssue == "no" && securedLoansInfo.monthlyNetIncome  > 15000 ){
-    // await addDoc(collection(db, "securedLoans"), {
-    //         ...securedLoansInfo,uId: userdata.uid,loanType:loanName,grade:"B"});
+   
     alert("grade B")
   }
   else if(securedLoansInfo.propertyStatus == "rented" && securedLoansInfo.CibilIssue == "yes" && securedLoansInfo.monthlyNetIncome  > 0 ){
-    // await addDoc(collection(db, "securedLoans"), {
-    //         ...securedLoansInfo,uId: userdata.uid,loanType:loanName,grade:"C"});
+   
     alert("grade C")
   }
   else {
