+++ src/Finance/Redux-Toolkit/slices/SecuredLoansCounter.js
@@ -368,7 +368,7 @@ export const SecuredLoansSlice = createSlice({
     },
     
   ],
-    securedLoansInfo:{};
+    securedLoansInfo:{}
   },
   reducers: {
